Move out of db.py
