+++ backend/pretix/__init__.py
@@ -1,5 +1,6 @@
 import logging
 import typing
+from dataclasses import dataclass
 from datetime import date
 from urllib.parse import urljoin
 
@@ -347,3 +348,34 @@ def create_order(conference: Conference, order_data: CreateOrderInput) -> Order:
     data = response.json()
 
     return Order(code=data["code"], payment_url=data["payments"][0]["payment_url"])
+
+
+@dataclass
+class Conference:
+    pretix_organizer_id: str
+    pretix_event_id: str
+
+
+def user_has_admission_ticket(
+    *, email: str, event_organizer: str, event_slug: int
+) -> bool:
+    response = pretix(
+        conference=Conference(
+            pretix_organizer_id=event_organizer, pretix_event_id=event_slug
+        ),
+        endpoint="tickets/attendee-has-ticket",
+        method="post",
+        json={
+            "attendee_email": email,
+            # TODO: In the future this method should be changed to send multiple events
+            "events": [
+                {
+                    "organizer_slug": event_organizer,
+                    "event_slug": event_slug,
+                }
+            ],
+        },
+    )
+    response.raise_for_status()
+    data = response.json()
+    return data["user_has_admission_ticket"]
